Move the calDistance function to common
So that we can start cleaning up this web of circular dependencies.
The main change is to tripManager.py and common.py.
All the others just fix dependencies.

All references to calDistance in tripManager are gone.

bash-3.2$ grep -r tripManager.*calDistance .
Binary file ./.git/.COMMIT_EDITMSG.swp matches
bash-3.2$

And the testCommon tests pass.

Ran 12 tests in 0.490s
OK
