+++ CFC_WebApp/main/LCS.py
@@ -1,5 +1,5 @@
 from __future__ import division
-from tripManager import calDistance
+from common import calDistance
 def lcs(a, b,radiusBound):
     lengths = [[0 for j in range(len(b)+1)] for i in range(len(a)+1)]
     # row 0 and column 0 are initialized to 0 already
