+++ CFC_WebApp/main/work_place_2.py
@@ -2,8 +2,7 @@ __author__ = 'Yin'
 from pymongo import MongoClient
 from home import detect_home, detect_home_from_db
 from home_2 import detect_home_2, detect_home_from_db_2
-from tripManager import calDistance
-from common import Is_weekday, get_static_pnts, most_common_2, calculate_appearance_rate, Is_date, Is_place_2
+from common import Is_weekday, get_static_pnts, most_common_2, calculate_appearance_rate, Is_date, Is_place_2, calDistance
 from dateutil import parser
 from get_database import get_section_db, get_profile_db
